Addressed code review comments
