+++ src/WebJobs.Script.WebHost/DependencyInjection/DependencyValidator/DependencyValidator.cs
@@ -29,7 +29,7 @@ namespace Microsoft.Azure.WebJobs.Script.WebHost.DependencyInjection
             expected.ExpectNone<IEventGenerator>();
 
             expected.Expect<ILoggerFactory, ScriptLoggerFactory>();
-            expected.ExpectFactory<IMetricsLogger, WebHostMetricsLoggerDecorator>();
+            expected.ExpectFactory<IMetricsLogger, NonDisposableMetricsLogger>();
 
             expected.Expect<IWebJobsExceptionHandler, WebScriptHostExceptionHandler>();
